Separated switch to Main class, implemented createRemoteDir()
